Update LoadDimensionOperator and DataQualityOperator
Bug fix: Update LoadDimensionOperator and DataQualityOperator

- Allow user to append data to Dimension tables in addition to just truncate-insert through LoadDimensionOperator
- Update DataQualityOperator to be more generic. Allowing user to pass sql check and expected result as parameter

+++ plugins/operators/load_dimension.py
@@ -10,6 +10,7 @@ class LoadDimensionOperator(BaseOperator):
     def __init__(self,
                  redshift_conn_id="",
                  destination_table="",
+                 append_data="",
                  sql_query="",
                  *args, **kwargs):
 
@@ -19,6 +20,8 @@ class LoadDimensionOperator(BaseOperator):
         Arguments: 
             redshift_conn_id   --  Redshift connection Id
             destination_table  --  Redshift table that need to be created
+            append_data        --  Allow user to switch between append and truncate-insert. 
+                                   If it's true, it will append data at the end of the table. Otherwise, it will truncate-insert
             sql_query          --  Dimension insert sql query
 
         Returns: 
@@ -28,6 +31,7 @@ class LoadDimensionOperator(BaseOperator):
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.destination_table = destination_table
+        self.append_data = append_data
         self.sql_query = sql_query
 
     def execute(self, context):
@@ -37,6 +41,8 @@ class LoadDimensionOperator(BaseOperator):
         Arguments: 
             redshift_conn_id   --  Redshift connection Id
             destination_table  --  Redshift table that need to be created
+            append_data        --  Allow user to switch between append and truncate-insert. 
+                                   If it's true, it will append data at the end of the table. Otherwise, it will truncate-insert
             sql_query          --  Dimension insert sql query
 
         Returns: 
@@ -45,9 +51,10 @@ class LoadDimensionOperator(BaseOperator):
 
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
         
-        self.log.info("Clearing data from destination Redshift table")
-        redshift.run("DELETE FROM {}".format(self.destination_table))
-        
-        self.log.info("Insert data to Fact table")
+        if self.append_data == False:
+            self.log.info("Clearing data from destination {} table".format(self.destination_table))
+            redshift.run("DELETE FROM {}".format(self.destination_table))
+            
+        self.log.info("Append data to {} table".format(self.destination_table))
         formatted_sql = self.sql_query.format(self.destination_table)
         redshift.run(formatted_sql)
